[refactoring] Introduced a "server" directory
