+++ .eslintrc.json
@@ -16,7 +16,7 @@
     "no-this-before-super": "warn",
     "no-undef": "warn",
     "no-unreachable": "warn",
-    "no-unused-vars": "warn",
+    "no-unused-vars": ["warn", { "args": "none" }],
     "constructor-super": "warn",
     "valid-typeof": "warn"
   }
