+++ lib/server/app.js
@@ -7,8 +7,8 @@ const bodyParser = require('body-parser');
 const config = require('painless-config');
 const express = require('express');
 const logger = require('morgan');
-const mockInsights = require('./providers/logger/mockInsights');
-const CrawlerFactory = require('./lib/crawlerFactory');
+const mockInsights = require('../../providers/logger/mockInsights');
+const CrawlerFactory = require('../crawlerFactory');
 const sendHelper = require('./middleware/sendHelper');
 
 auth.initialize(config.get('CRAWLER_SERVICE_AUTH_TOKEN') || 'secret', config.get('CRAWLER_SERVICE_FORCE_AUTH'));
