+++ lib/server/middleware/auth.js
