+++ lib/server/routes/config.js
