+++ lib/server/routes/requests.js
@@ -4,8 +4,8 @@
 const auth = require('../middleware/auth');
 const express = require('express');
 const expressJoi = require('express-joi');
-const Request = require('../lib/request');
-const TraversalPolicy = require('../lib/traversalPolicy');
+const Request = require('../../request');
+const TraversalPolicy = require('../../traversalPolicy');
 const wrap = require('../middleware/promiseWrap');
 
 const requestsSchema = {
