+++ lib/server/routes/status.js
