+++ lib/server/routes/webhook.js
@@ -4,7 +4,7 @@
 const crypto = require('crypto');
 const express = require('express');
 const moment = require('moment');
-const Request = require('../lib/request');
+const Request = require('../../request');
 const wrap = require('../middleware/promiseWrap');
 
 let crawlerService = null;
